Separated model classes into one file per class
